Connection

+++ index.php
@@ -24,37 +24,40 @@ session_start();
   <?php
   header ('Content-Type: text/html; charset=utf-8');
   require('Inc/require.inc.php');
+  if (!isset($_SESSION)) { session_start(); }
   //require('Inc/globals.inc.php');
   $EX = isset($_REQUEST['EX']) ? $_REQUEST['EX'] : 'home';
-  if(isset($_REQUEST['idPrev'])){
-      $idPrev= $_REQUEST['idPrev'];
-      $idNext= $_REQUEST['idNext'];
+  if(!isset($_SESSION['USER'])){
+    home();
   }
-  switch($EX)
-  {
-      case 'home'      : home();       break;
-      case 'searchFacility'      : searchFacility();       break;
-      case 'createFacility'      : createFacility();       break;
-      case 'deleteFacility'      : deleteFacility();       break;
-      case 'searchFighter'      : searchFighter();       break;
-      case 'createFighter'      : createFighter();       break;
-      case 'deleteFighter'      : deleteFighter();       break;
-      case 'searchItem'      : searchItem();       break;
-      case 'createItem'      : createItem();       break;
-      case 'deleteItem'      : deleteItem();       break;
-      case 'searchPlayer'      : searchPlayer();       break;
-      case 'deletePlayer'      : deletePlayer();       break;
-      case 'searchCenter'      : searchCenter();       break;
-      case 'createCenter'      : createCenter();       break;
-      case 'deleteCenter'      : deleteCenter();       break;
-      case 'searchClub'      : searchClub();       break;
-      case 'createClub'      : createClub();       break;
-      case 'deleteClub'      : deleteClub();       break;
-      case 'searchCompetition'      : searchCompetition();       break;
-      case 'createCompetition'      : createCompetition();       break;
-      case 'deleteCompetition'      : deleteCompetition();       break;
-      case 'admin'      : admin();       break;
-      default : home();
+  else{
+    switch($EX)
+      {
+          case 'home'      : home();       break;
+          case 'searchFacility'      : searchFacility();       break;
+          case 'createFacility'      : createFacility();       break;
+          case 'deleteFacility'      : deleteFacility();       break;
+          case 'searchFighter'      : searchFighter();       break;
+          case 'createFighter'      : createFighter();       break;
+          case 'deleteFighter'      : deleteFighter();       break;
+          case 'searchItem'      : searchItem();       break;
+          case 'createItem'      : createItem();       break;
+          case 'deleteItem'      : deleteItem();       break;
+          case 'searchPlayer'      : searchPlayer();       break;
+          case 'deletePlayer'      : deletePlayer();       break;
+          case 'searchCenter'      : searchCenter();       break;
+          case 'createCenter'      : createCenter();       break;
+          case 'deleteCenter'      : deleteCenter();       break;
+          case 'searchClub'      : searchClub();       break;
+          case 'createClub'      : createClub();       break;
+          case 'deleteClub'      : deleteClub();       break;
+          case 'searchCompetition'      : searchCompetition();       break;
+          case 'createCompetition'      : createCompetition();       break;
+          case 'deleteCompetition'      : deleteCompetition();       break;
+          case 'admin'      : admin();       break;
+          case 'deconnection'      : deconnection();       break;
+          default : home();
+      }
   }
   require('./View/layout.view.php');
   function home()
@@ -240,5 +243,16 @@ session_start();
       $page['class'] = 'VHtml';
       $page['method'] = 'showHtml';
       $page['arg'] = './Html/admin.php';
+  }
+
+  function deconnection()
+  {
+    global $page;
+    unset($_SESSION['USER']);
+    session_destroy();
+    $page['title'] = 'Retour après déco';
+    $page['class'] = 'VHtml';
+    $page['method'] = 'showHtml';
+    $page['arg'] = 'Html/accueil.php';
   }
   ?>
